refactor componetize
